+++ src/services/job/api-path.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-anonymous-default-export */
-const JOB = "http://103.20.97.112:8000/api/v1/jobs";
+const JOB = "http://103.20.97.112:8000/ai-service/api/v1/jobs";
 
 const JOB_DOCUMENT_HISTORY = `${JOB}/documents/history`;
